+++ lib/redis_failover/client.rb
@@ -207,13 +207,7 @@ module RedisFailover
       verify_supported!(method)
       tries = 0
       begin
-        if REDIS_READ_OPS.include?(method)
-          # send read operations to a slave
-          slave.send(method, *args, &block)
-        else
-          # direct everything else to master
-          master.send(method, *args, &block)
-        end
+        client_for(method).send(method, *args, &block)
       rescue *CONNECTIVITY_ERRORS => ex
         logger.error("Error while handling `#{method}` - #{ex.inspect}")
         logger.error(ex.backtrace.join("\n"))
@@ -225,6 +219,12 @@ module RedisFailover
           retry
         end
         raise
+      ensure
+        if info = Thread.current[:last_operation_info]
+          if info[:method] == method
+            Thread.current[:last_operation_info] = nil
+          end
+        end
       end
     end
 
@@ -400,6 +400,34 @@ module RedisFailover
     def recently_heard_from_node_manager?
       return false unless @last_znode_timestamp
       Time.now - @last_znode_timestamp <= ZNODE_UPDATE_TIMEOUT
+    end
+
+    # Returns the client to use for the specified operation.
+    #
+    # @param [Symbol] method the method for which to retrieve a client
+    # @return [Redis] a redis client to use
+    # @note
+    #   This method stores the last client/method used to handle the case
+    #   where the same RedisFailover::Client instance is referenced by a
+    #   block passed to multi.
+    def client_for(method)
+      if info = Thread.current[:last_operation_info]
+        return info[:client]
+      elsif REDIS_READ_OPS.include?(method)
+        # send read operations to a slave
+        Thread.current[:last_operation_info] = {
+          :client => slave,
+          :method => method
+        }
+      else
+        # direct everything else to master
+        Thread.current[:last_operation_info] = {
+          :client => master,
+          :method => method
+        }
+      end
+
+      Thread.current[:last_operation_info][:client]
     end
   end
 end