+++ Changes.md
@@ -2,6 +2,7 @@ HEAD
 -----------
 - Added a way to gracefully shutdown a RedisFailover::Client. (#13)
 - Upgraded to latest ZK version that supports forking.
+- Handle case where the same RedisFailover::Client is referenced by a #multi block (#14)
 
 0.8.2
 -----------